requirements.txt
# Kaggle Tabular Agent - Requirements

# LLM
llama-cpp-python>=0.2.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
lightgbm>=4.0.0
xgboost>=2.0.0
catboost>=1.2.0
scikit-learn>=1.3.0

# Kaggle API
kaggle>=1.6.0

# Utilities
tqdm>=4.65.0
